requirements.txt
birdnetlib
numpy
pyyaml
librosa
tensorflow; platform_system != "Darwin"
tensorflow-macos; platform_system == "Darwin"
pygbif
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
databases[sqlite]
pydantic
httpx
astral
pytest

